btrfs: fix build with fresh linux/btrfs.h
Fix build with fresh <linux/btrfs.h> that provides a slightly different
definition of struct btrfs_ioctl_defrag_range_args.

* btrfs.c (btrfs_ioctl): Cast "len" member of
struct btrfs_ioctl_defrag_range_args to uint64_t.
* tests/btrfs.c (btrfs_print_defrag_range_args): Cast "start" and "len"
members of struct btrfs_ioctl_defrag_range_args to uint64_t.

+++ btrfs.c
@@ -604,7 +604,7 @@ MPERS_PRINTER_DECL(int, btrfs_ioctl,
 
 		tprintf("{start=%" PRIu64 ", len=", (uint64_t)args.start);
 
-		tprintf("%" PRIu64, args.len);
+		tprintf("%" PRIu64, (uint64_t) args.len);
 		if (args.len == UINT64_MAX)
 			tprints(" /* UINT64_MAX */");
